Changed the workflow on how training programs are being ran. With this proposal after clicking the "done" icon, you will first be prompted your current progress and then how you'd like to change your values for the next time.
For this I piped the dismiss callback from the dialog down one
form the show edit dialog function to allow me to present a succesion of
dialogs to the user. Additionaly I used the setTitle function from the
previous commit.

Code quality might not yet be on par since I only quickly added this for
personal use.